pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-ticket-router-env"
version = "1.0.0"
description = "TicketRouterEnv — SaaS ticket routing simulator for OpenEnv"
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "openai>=1.40.0",
    "websockets>=12.0",
    "pyyaml>=6.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
include = ["env*", "server*"]
